chore(lint): clippy cleanup — 321 -> 60 warnings, crate-level allows for arg-count/type-width
Mechanical lint fixes across 63 files (three parallel review agents,
disjoint file buckets): collapsible ifs, redundant field names, needless
borrows, is_multiple_of/div_ceil, sort_by_key, redundant closures,
same-type casts, slice-size calcs, doc-list indentation, integer-only
clamp conversions, and excessive-precision float literals verified
bit-identical (EMBED_SCALE, GELU_TANH_COEF hex-equal).

Deliberately NOT fixed (~45, equivalence not provable mechanically):
float max/min chains (clamp changes NaN semantics), loop-index lints
where the index feeds offset arithmetic or multiple arrays (CPU kernel
oracles), a clippy unwrap_or suggestion that would change flag parsing
(64|128 match guard), signature-changing to_*-by-value suggestions.

too_many_arguments + type_complexity allowed crate-wide (user sign-off:
kernel dispatch plumbing's scalar dims are the domain, not a smell);
12 per-item allows removed as redundant.

Gates: golden 8/8 byte-identical post-cleanup; suite 596/0; build clean.

+++ src/dgq/fp4.rs
@@ -73,7 +73,7 @@ pub fn fp8_e4m3_from_f32(f: f32) -> u8 {
         let adjusted_bits = adjusted.to_bits();
         (adjusted_bits - DENORM_MASK) as u8
     } else {
-        let mant_odd = ((f_bits >> 20) & 1) as u32;
+        let mant_odd = (f_bits >> 20) & 1;
         f_bits = f_bits.wrapping_add((7u32.wrapping_sub(127) << 23) + 0x7_FFFF);
         f_bits = f_bits.wrapping_add(mant_odd);
         (f_bits >> 20) as u8
